log: simpler, more flexible interface for context logging
accepting just the 'module' will not be very helpful for comprehensive logging;
in an of itself this info is just not that helpful. 'module' is just one of many
fields that will have to be tracked across layers, and not the most important one.

the package interface is therefore changed so that a logger can accept
an arbitrary 'context', with the intention of propagating it to lower layers
for further enrichment.

any component can implement the ContextLogger interface to support inheriting context
from upper layers.

the Logger type is a very thin wrapper over logrus.Logger (in fact - logrus.Entry
which can be initialized with some context ootb), and the interface is slightly
less verbose than 'logrus.WithFields(logrus.Fields{...' etc.

Issues: MEN-532

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>