config: introduce global configuration available for all components on demand
this is in preparation for a refactoring which will simplify log context passing.

the idea is to make the components like devauth, datastore, etc. short-lived,
instantiated on demand in the handler, and able to accept a log with the *current* context
via dependency injection (property-based DI). to enable this, the config must be exposed
somehow, instead of being 'swallowed' by main().

having them created at runtime as singletons, with embedded loggers, is not friendly to
structured logging, where current context is key. this could be worked around with e.g.
extending the component interfaces with a WithLog() method, but this is clumsy and verbose;

since devauth is still a pretty small service, a large refact like this is affordable.

Issues: MEN-532

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>