client_inventory: inventory service client
Minimal client for inventory service that allows for creating a device.
There is some special handling of inventory service response status
codes. We normally expect `201 Created` in a response to POST, as the
device was just created. However, a special case is when a device
already exists. In this case the server will respond with `409
Conflict`, but we still treat is as success.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>