devauth: handle case when device is in DB but no authorization requests were stored
Handle an endge case when a device is present in DB, but there are no
registered authorization requests. In this case, the preliminary checks
returned a device, no errors yet, then we queried for registered
authorization requests, but that returned an empty slice. With Go magic,
empty slice is not nil, but nil slice is empty (len() for both is 0).

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>