tests: acceptance: update acceptance tests with authentication data sets and identity split
Update tests to match changes that were introduced as part of identity split.
The tests use an updated API.

Most changes are centered around the fact that device ID is now assigned by the
system and cannot be determined upfront. For this reason, the tests will try to
find a matching device in the listing of all devices. Once this is done, the
tests will try to find the ID of authentication data set and act on this instead
of on a device.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>