devauth: reject tokens if device is being decommissioned
The fix addresses a race condition in device decommissioning.

What happens is that when decommissioning starts we mark a device as being
removed (model.Device.Decommissioning is true). This is to handle a situation
when a device being removed sends an auth request, in which case the request is
denied. Device data (its tokens and such) is removed only once we have
successfully submitted device removal job to the orchestrator. However, token
verification does not check if device is being removed and will accept its
token.

The race condition enables a situation in which device request (eg. inventory
attributes update) is allowed to pass through the gateway. Depending on the
order of events (eg. processing of removal of device inventory data wrt. device
inventory attributes update) we may end up with device request is processed
after device data DELETE request. Depending on a service, the end result may be
that the device data still exists in the system after it has been
decommissioned.

Fixes: [Mender #1210]
Changelog: None
Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>