devauth, store/mongo, model: introduce DeviceUpdate wrapper
DataStoreMongo.DeviceUpdate() operation was not updated when decommissioning was
introduced. For this reason, model.Device.Decommissioning field was never a part
of Mongo update and its status was not recorded in DB. The patch addresses this
problem by introducing model.DeviceUpdate helper. This allows to remove
makeUpdate() helper and makes update device operation to be in line with
UpdateAuthSet() call.

As a side effect, failure to record model.Device.Decommissioning status resulted
in device tokens being accepted even though it was being decommissioned.

Fixes: [Mender #1210]
Changelog: none
Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>