tests: replace conductor with mock API implementation
Testing with conductor is unstable in the sense that conductor will sometimes
throw 500 in response to rqeuests, or just fail to talk to dynomite just because
the tests are run in specific way or happen too fast. This would cause the tests
to fail and ideally these problems should be caught in integration.

The change removes conductor from compose file (as a side effect the tests run
faster now).

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>

Changelog: None