docs: move preauth endpoint from internal to management
it's a mixup - the api was supposed to be public according
to the task description.

the reason was that we're treating deviceadm as an external service,
possibly provided by the user. we made similar provisions already
in other endpoints, so we continue with this one too.

Issues: MEN-1630

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>