devauth: PreauthorizeDevice method
factoring this out as a clean, separate func. there are some common
parts with SubmitAuthRequest, but the latter is in general a mess riddled
with usecase-specific logic.

there's a gotcha regarding mongo:
- we want to reject conflicting id data
- it's simple on the Device level (unique index on id data)
- to be completely correct, we should also have the same check when inserting an auth set
- but it's not possible - we can't have an index on id data alone, would disallow key rotation
- tried work around with some upsert/setOnInsert tricks, but in the end this also requires a unique index

so, we only signal conflict when there is a Device with this id data. the important point
is that the 'delete auth set' method should clean up the Device when the last auth set is removed.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>