tests: adjust mock tenantadm's response
the tenantadm client, when it gets 401 on token verify,
tries to parse out the error message. to do this, it expects
a well formed mender error response (request_id + error).

it's irrelevant if the message is filled in, but the structure
must be correct - so adjust the fake tenantadm to return one.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>