tests: skip preauth 'conflict' tests
has to wait until iddata normalization is merged in POST auth_requests.

we use it in fixtures to prepare devices, but id data is still saved verbatim
(python encoding).

when trying to test conflict we get 200, because the preauth ep already normalizes
id data and they're recognized as new auth sets.

not fixing this since those changes are already on master.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>