devauth: fix context handling in ProvisionTenant
MigrateTenant always migrates the default db.

the main cause is - we're not setting up the context correctly with the
tenant's id.

actually, it works in part on the default db, in part on the tenant's db
(we use straight dbname in places, and context in others).

causes extremely difficult bugs when e.g. migrating from os to
enterprise (tenant's db needs migrating, so we try - but data
is taken from the default db, so it fails on dup keys).

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>