client/tenantadm: return full Tenant
have an api specific struct and return it from the
verification method.

on tenantadm's side, the tenant is already returned so no
modifications there.

we actually need the Plan of the tenant when issuing the token -
but return the full struct, maybe sth else will come in handy.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>