devauth: insert tenant's Plan into JWT
upon an auth request, device will obtain a JWT
with tenant's Plan field.

this is a mirror mechanism of user login.

the JWT can then be inspected by other services to switch
features based on plan.

some refact was done here because the current logic
was not super easy to work with and slightly confusing
(esp. around default tenant token handling). the new
flow is hopefully simpler and more descriptive of what
it actually going on.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>