cache: redis cache with throttling and token mgmt
the cache applies throttling and fetches the cached token in one go.
this results in just a single network roundtrip to redis - both
operations are wrapped in a pipeline (in fact, a transactional
pipeline).

internally, the simplest ratelimiting redis patterns are used (see
comments).

miniredis is a full in process mock of the redis server, seems widely
used so I gave it a try for testing. checked that it's extremely fast,
even hundreds of test loops and re-runs pass in ~1 sec.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>