cache: honor burst interval when throttling device
the device may be locked out of the api forever if it keeps
calling with an interval smaller than min_interval_sec of the burst.
this is because there's just a single burst key tracking the usage,
with a TTL pushed forward with every request.

instead, we want the device to succeed every min_interval_sec,
regardless how quick it is, so:
- track the bursts in 'windowed' burst keys (one for every
min_interval_sec window of time - just like with quotas)
- allow them to expire in min_interval_sec

every min_interval_sec, the current window number will roll over, and
the device won't hit the new/nonexistent burst key. it will be created in the same
instance to prevent further requests, but this one call will succeed.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>