cache_test: general (timing) improvements
for existing tests, move time in sync both in miniredis and
Cache instance.

this is the right way of testing limits - the time must move
for the Cache as well, because it is responsible for computing
the interval/window sequence numbers for both quotas and bursts.

otherwise only the simplest scenarios can be tested (e.g. no
rollover of current interval).

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>