fix: Inconsistent device `check_in_time` when listing devices
The lookup for `check_in_time` from the cache does not work when running
Redis in cluster mode because of the MGET command requires keys to hash
to the same slot. This commit replaces MGET with multiple batched GET
commands when running Redis in cluster mode.

Changelog: Commit
Ticket: MEN-7337
Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>